+++ components/Navbar.js
@@ -7,11 +7,10 @@ export default function Navbar() {
     <>
         <div className='navbar'>
             <Link href="/"><a>Home</a></Link>
-            {/* <Link href="/login"><a>Login</a></Link> */}
+            <Link href="/login"><a>Login</a></Link>
             <Link href="/register"><a>Register</a></Link>
             <Link href="/profile"><a>Profile</a></Link>
             <Link href="/users"><a>Users</a></Link>
-            <Link href="/test"><a>Test CSS</a></Link>
         </div>
     </>
   )
