+++ pages/index.js
@@ -9,35 +9,6 @@ import { setCookie, getCookie, getCookies, deleteCookie } from "cookies-next"
 export default function Home({ users }) {
     const router = useRouter()
 
-    const [loginDataForm, setLoginDataForm] = useState({
-        username: '',
-        password: '',
-    })
-
-    const handleChange = (e) => {
-        const value = e.target.value;
-        setLoginDataForm({
-          ...loginDataForm,
-          [e.target.name]: value
-        });
-    }
-
-    const loginSubmit = (e) => {
-        e.preventDefault()
-        const {username, password} = loginDataForm
-        apiService.login(loginDataForm)
-        .then(response => {
-            let tokenData = {
-                refreshToken: response.data.refreshToken
-            }
-            apiService.post('tokens',tokenData).then(response => setCookie('loginSuccessfull', true))
-            return router.push('/profile')
-        })
-        .catch(error => {
-            // console.log('Catch in index loginSubmit: ', error.response.message)
-        })
-    }
-
     return (
         <>
             <Head>
@@ -47,29 +18,10 @@ export default function Home({ users }) {
             </Head>
             <Navbar />
             <div className='container'>
-                <h1 className='register'>Login</h1>
-                <form className='formLogin' onSubmit={loginSubmit} method="post">
-
-                <label htmlFor="username">Username:</label>
-                <input type="text" id="username" name="username" value={loginDataForm.value} onChange={handleChange} />
-
-                <label htmlFor="password">Password:</label>
-                <input type="password" id="password" name="password" value={loginDataForm.value} onChange={handleChange} />
-
-                <button type="submit">Login</button>
-                </form>
+                <h1>Home</h1>
             </div>
         </>
     )
 }
 
-export async function getStaticProps() {
-    const res = await fetch('http://localhost:5000/api/users')
-    const users = await res.json()
 
-    return {
-        props: {
-            users,
-        },
-    }
-}
